+++ package/etc/conf.d/conflib/vendor_juniper/junos.conf
@@ -1,8 +1,7 @@
-filter f_juniper_junos {
+filter f_juniper_junos_structured {
     match('^\[junos@2636' value("SDATA"))
-        or match('syslog@juniper.net' value("SDATA"))
 };
 
-filter f_juniper_junos_legacy {
+filter f_juniper_junos_standard {
     program("RT_IDP|RT_FLOW|RT_IDS|RT_UTM|Juniper");
 };
