+++ package/etc/conf.d/conflib/vendor_juniper/legacy.conf
@@ -14,6 +14,7 @@ filter f_juniper_netscreen {
 
 };
 
-filter f_juniper_idp {
-    match("juniper_idp", value("fields.sc4s_vendor_product") type(glob) );
+ filter f_juniper_idp {
+    match("juniper_idp", value("fields.sc4s_vendor_product") type(glob))
+      or match('^\[syslog@juniper' value("SDATA"))
 };
