+++ package/etc/conf.d/log_paths/p_rfc3164-juniper-idp.conf
@@ -1,7 +1,7 @@
 log {
 
     source(s_default-ports);
-    filter(f_is_rfc3164);
+    filter(f_is_rfc5424_strict);
     filter(f_juniper_idp);
 
     rewrite { r_set_splunk_dest_default(sourcetype("juniper:idp"), index("netids"))};
@@ -9,7 +9,7 @@ log {
                 p_add_context_splunk(key("juniper_idp"));
     };
     rewrite {
-        r_set_splunk(template("t_standard"))
+        r_set_splunk(template("t_hdr_sdata_msg"))
     };
     destination(d_hec);  #--HEC--
     flags(flow-control);
