+++ package/etc/conf.d/log_paths/p_rfc3164-juniper-junos.conf
@@ -6,7 +6,7 @@ log {
 
     source(s_default-ports);
     filter(f_is_rfc3164);
-    filter(f_juniper_junos_legacy);
+    filter(f_juniper_junos_standard);
 
     if (program('RT_IDP')) {
         rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:idp"), index("netids"))};
@@ -15,7 +15,7 @@ log {
         rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall"), index("netfw"))};
         parser {p_add_context_splunk(key("juniper_junos_flow")); };
     } elif (program('RT_IDS')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall"), index("netids"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:idp"), index("netids"))};
         parser {p_add_context_splunk(key("juniper_junos_ids")); };
     } elif (program('RT_UTM')) {
         rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall"), index("netids"))};
