+++ package/etc/conf.d/log_paths/p_rfc_5424_strict-juniper-junos.conf
@@ -6,30 +6,33 @@ log {
 
     source(s_default-ports);
     filter(f_is_rfc5424_strict);
-    filter(f_juniper_junos);
+    filter(f_juniper_junos_structured);
 
     if (program('RT_IDP')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:idp:structured"), index("netids"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:idp:structured"), index("netids")) };
         parser {p_add_context_splunk(key("juniper_idp")); };
     } elif (program('RT_FLOW')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netfw"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netfw")) };
         parser {p_add_context_splunk(key("juniper_junos_flow")); };
     } elif (program('RT_IDS')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netids"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netids")) };
         parser {p_add_context_splunk(key("juniper_junos_ids")); };
     } elif (program('RT_UTM')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netids"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:junos:firewall:structured"), index("netids")) };
         parser {p_add_context_splunk(key("juniper_junos_utm")); };
-    } elif (program('Jnpr')) {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:idp:structured"), index("netids"))};
-        parser {p_add_context_splunk(key("juniper_junos_idp")); };
-        }
+    }
+# Legacy Netscreen IDP is handled in the "p_rfc3164-juniper-idp.conf" log path
+#
+#    } elif (program('Jnpr')) {
+#        rewrite { r_set_splunk_dest_default(sourcetype("juniper:idp:structured"), index("netids")) };
+#        parser {p_add_context_splunk(key("juniper_junos_idp")); };
+#        }
     else {
-        rewrite { r_set_splunk_dest_default(sourcetype("juniper:structured"), index("netops"))};
+        rewrite { r_set_splunk_dest_default(sourcetype("juniper:structured"), index("netops")) };
         parser {p_add_context_splunk(key("juniper_structured")); };
     };
 
-    rewrite { r_set_splunk(template("t_hdr_sdata_msg"))};
+    rewrite { r_set_splunk(template("t_hdr_sdata_msg")) };
     destination(d_hec);  #--HEC--
