+++ package/etc/context-local/splunk_index.csv
@@ -9,10 +9,11 @@
 #juniper_idp,index,netids
 #juniper_junos_flow,index,netfw
 #juniper_junos_ids,index,netids
-#juniper_junos_utm,index,netids
+#juniper_junos_utm,index,netfw
 #juniper_sslvpn,index,netfw
 #juniper_netscreen,index,netfw
-#juniper_nsm,index,netids
+#juniper_nsm,index,netfw
+#juniper_nsm_idp,index,netids
 #juniper_legacy,index,netops
 #pan_traffic,index,netfw
 #pan_threat,index,netproxy
